@@ -1,10 +1,10 @@
 name: flutter_widget_from_html_core
-version: 0.15.2
+version: 0.16.0
 description: Flutter package to render html as widgets that focuses on correctness and extensibility.
 homepage: https://github.com/daohoangson/flutter_widget_from_html/tree/master/packages/core
 
 environment:
-  flutter: ">=3.22.0"
+  flutter: ">=3.32.0"
   sdk: ">=3.4.0 <4.0.0"
 
 dependencies:
