@@ -1,7 +1,3 @@
-// TODO: remove lint ignore when our minimum Flutter version >= 3.19
-// ignore: unnecessary_import
-import 'dart:ui';
-
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 
@@ -33,7 +29,7 @@ class HtmlListMarker extends LeafRenderObjectWidget {
       : markerType = HtmlListMarkerType.square;
 
   @override
-  RenderObject createRenderObject(BuildContext _) =>
+  RenderObject createRenderObject(BuildContext context) =>
       _ListMarkerRenderObject(markerType, textStyle);
 
   @override
@@ -44,7 +40,7 @@ class HtmlListMarker extends LeafRenderObjectWidget {
   }
 
   @override
-  void updateRenderObject(BuildContext _, RenderObject renderObject) {
+  void updateRenderObject(BuildContext context, RenderObject renderObject) {
     (renderObject as _ListMarkerRenderObject)
       ..markerType = markerType
       ..textStyle = textStyle;
