@@ -607,15 +607,16 @@ class Explainer {
       return _widget(widget.builder(context));
     }
 
-    if (widget is LayoutBuilder) {
+    if (widget is LayoutBuilder ||
+        widget.runtimeType.toString() == 'HtmlLayoutBuilder') {
       return _widget(
-        widget.builder(
+        (widget as dynamic).builder(
           context,
           BoxConstraints.loose(
             TestWidgetsFlutterBinding
                 .instance.platformDispatcher.implicitView!.physicalSize,
           ),
-        ),
+        ) as Widget,
       );
     }
 
@@ -897,7 +898,7 @@ class HitTestApp extends StatelessWidget {
   const HitTestApp({required this.html, super.key, required this.list});
 
   @override
-  Widget build(BuildContext _) => MaterialApp(
+  Widget build(BuildContext context) => MaterialApp(
         home: Scaffold(
           body: HtmlWidget(
             html,
