@@ -418,57 +418,6 @@ Future<void> main() async {
     });
   });
 
-  group('width', () {
-    testWidgets('renders without width', (WidgetTester tester) async {
-      const html = '<table><tr><td>Foo</td></tr></table>';
-      final e = await explain(tester, html, useExplainer: false);
-      expect(e, contains('└HtmlTableCell(columnStart: 0, rowStart: 0)'));
-    });
-
-    testWidgets('renders width: 50px', (WidgetTester tester) async {
-      const html = '<table><tr><td style="width: 50px">Foo</td></tr></table>';
-      final explained = await explain(tester, html, useExplainer: false);
-      expect(
-        explained,
-        isNot(contains('└HtmlTableCell(columnStart: 0, rowStart: 0)')),
-      );
-      expect(
-        explained,
-        contains('└HtmlTableCell(columnStart: 0, rowStart: 0, width: 50.0)'),
-      );
-    });
-
-    testWidgets('renders width: 100%', (WidgetTester tester) async {
-      const html = '<table><tr><td style="width: 100%">Foo</td></tr></table>';
-      final explained = await explain(tester, html, useExplainer: false);
-      expect(
-        explained,
-        isNot(contains('└HtmlTableCell(columnStart: 0, rowStart: 0)')),
-      );
-      expect(
-        explained,
-        contains('└HtmlTableCell(columnStart: 0, rowStart: 0, width: 100.0%)'),
-      );
-    });
-
-    testWidgets('renders width: 100% within TABLE', (tester) async {
-      const html = '<table><tr><td>'
-          '<table><tr><td style="width: 100%">'
-          'Foo'
-          '</td></tr></table>'
-          '</td></tr></table>';
-      final explained = await explain(tester, html, useExplainer: false);
-      expect(
-        explained,
-        contains('└HtmlTableCell(columnStart: 0, rowStart: 0)'),
-      );
-      expect(
-        explained,
-        contains('└HtmlTableCell(columnStart: 0, rowStart: 0, width: 100.0%)'),
-      );
-    });
-  });
-
   group('combos', () {
     testWidgets('renders nested table', (WidgetTester tester) async {
       // https://github.com/daohoangson/flutter_widget_from_html/issues/1070
@@ -494,7 +443,7 @@ Future<void> main() async {
       const windowSize = 100.0;
       tester.setWindowSize(const Size(windowSize, windowSize));
 
-      const html = '<table><tr><td align="center">'
+      const html = '<table style="width: 100%"><tr><td align="center">'
           '<table><tr><td>Foo</td></tr></table>'
           '</td></tr></table>';
       await explain(tester, html);
@@ -653,81 +602,6 @@ Future<void> main() async {
     });
   });
 
-  group('background', () {
-    testWidgets('cell color', (WidgetTester tester) async {
-      // https://github.com/daohoangson/flutter_widget_from_html/issues/171
-      const html = '<table><tr>'
-          '<td style="background-color: #f00">Foo</td>'
-          '</tr></table>';
-      final explained = await explain(tester, html);
-      expect(
-        explained,
-        equals(
-          '[SingleChildScrollView:child=[HtmlTable:children='
-          '[HtmlTableCell:child='
-          '[Container:color=#FFFF0000,child='
-          '[Padding:(1,1,1,1),child='
-          '[Align:alignment=centerLeft,widthFactor=1.0,child='
-          '[RichText:(:Foo)]'
-          ']]]]]]',
-        ),
-      );
-    });
-
-    testWidgets('row color', (WidgetTester tester) async {
-      // https://github.com/daohoangson/flutter_widget_from_html/issues/1028
-      const html = '<table><tr style="background-color: #f00">'
-          '<td>Foo</td><td>Bar</td>'
-          '</tr></table>';
-      final explained = await explain(tester, html);
-      expect(
-        explained,
-        equals(
-          '[SingleChildScrollView:child=[HtmlTable:children='
-          '[HtmlTableCell:child='
-          '[Container:color=#FFFF0000,child='
-          '[Padding:(1,1,1,1),child='
-          '[Align:alignment=centerLeft,widthFactor=1.0,child='
-          '[RichText:(:Foo)]'
-          ']]]],'
-          '[HtmlTableCell:child='
-          '[Container:color=#FFFF0000,child='
-          '[Padding:(1,1,1,1),child='
-          '[Align:alignment=centerLeft,widthFactor=1.0,child='
-          '[RichText:(:Bar)]'
-          ']]]]'
-          ']]',
-        ),
-      );
-    });
-
-    testWidgets('overwrites row color', (WidgetTester tester) async {
-      const html = '<table><tr style="background-color: #f00">'
-          '<td>Foo</td><td style="background-color: #0f0">Bar</td>'
-          '</tr></table>';
-      final explained = await explain(tester, html);
-      expect(
-        explained,
-        equals(
-          '[SingleChildScrollView:child=[HtmlTable:children='
-          '[HtmlTableCell:child='
-          '[Container:color=#FFFF0000,child='
-          '[Padding:(1,1,1,1),child='
-          '[Align:alignment=centerLeft,widthFactor=1.0,child='
-          '[RichText:(:Foo)]'
-          ']]]],'
-          '[HtmlTableCell:child='
-          '[Container:color=#FF00FF00,child='
-          '[Padding:(1,1,1,1),child='
-          '[Align:alignment=centerLeft,widthFactor=1.0,child='
-          '[RichText:(:Bar)]'
-          ']]]]'
-          ']]',
-        ),
-      );
-    });
-  });
-
   testWidgets('renders display: table', (WidgetTester tester) async {
     const html = '''
 <div style="display: table">
@@ -747,10 +621,10 @@ Future<void> main() async {
       equals(
         '[SingleChildScrollView:child=[HtmlTable:children='
         '[HtmlTableCaption:child=[RichText:align=center,(:Caption)]],'
-        '[HtmlTableCell:child=[RichText:(+b:Header 1)]],'
-        '[HtmlTableCell:child=[RichText:(+b:Header 2)]],'
-        '[HtmlTableCell:child=[RichText:(:Value 1)]],'
-        '[HtmlTableCell:child=[RichText:(:Value 2)]]'
+        '[HtmlTableCell:child=[CssBlock:child=[RichText:(+b:Header 1)]]],'
+        '[HtmlTableCell:child=[CssBlock:child=[RichText:(+b:Header 2)]]],'
+        '[HtmlTableCell:child=[CssBlock:child=[RichText:(:Value 1)]]],'
+        '[HtmlTableCell:child=[CssBlock:child=[RichText:(:Value 2)]]]'
         ']]',
       ),
     );
@@ -767,8 +641,8 @@ Future<void> main() async {
       explained,
       equals(
         '[SingleChildScrollView:child=[HtmlTable:children='
-        '[HtmlTableCell:child=[RichText:(:Foo)]],'
-        '[HtmlTableCell:child=[RichText:(:Bar)]]'
+        '[HtmlTableCell:child=[CssBlock:child=[RichText:(:Foo)]]],'
+        '[HtmlTableCell:child=[CssBlock:child=[RichText:(:Bar)]]]'
         ']]',
       ),
     );
@@ -810,22 +684,6 @@ Future<void> main() async {
         await explain(t, '<table cellspacing="20"><tr><td>2</td></tr></table>');
         final after = t.table;
         expect(after.borderSpacing, equals(20.0));
-      });
-
-      testWidgets('updates maxWidths', (WidgetTester tester) async {
-        await explain(
-          tester,
-          '<div style="max-width: 100px"><table><tr><td>Foo</td></tr></table></div>',
-        );
-        final before = tester.table;
-        expect(before.maxWidth, equals(100.0));
-
-        await explain(
-          tester,
-          '<div style="max-width: 200px"><table><tr><td>Foo</td></tr></table></div>',
-        );
-        final after = tester.table;
-        expect(after.maxWidth, equals(200.0));
       });
 
       testWidgets('updates textDirection', (WidgetTester tester) async {
@@ -1115,14 +973,6 @@ Future<void> main() async {
         ]);
 
         expect(second.width, equals(first.width));
-
-        final m = _loggerMessages;
-        expect(m, contains(contains('Got child#0 min width:')));
-        expect(m, contains(contains('Got child#1 min width:')));
-        expect(m, contains(contains('Got child#2 size without contraints:')));
-        expect(m, isNot(contains(contains('Got child#2 min width:'))));
-        expect(m, contains(contains('Got child#3 size without contraints:')));
-        expect(m, isNot(contains(contains('Got child#3 min width:'))));
       });
     });
 
@@ -1168,12 +1018,29 @@ Future<void> main() async {
   <tr><td colspan="2">Lorem ipsum dolor sit amet.</td></tr>
   <tr><td>Foo</td><td>Bar</td></tr>
 </table>''',
-              'height_1px': 'Above<table border="1" style="height: 1px">'
-                  '<tr><td style="height: 1px">Foo</td></tr></table>Below',
               'rowspan': '''
 <table border="1">
   <tr><td rowspan="2">$multiline</td><td>Foo</td></tr>
   <tr><td>Bar</td></tr>
+</table>''',
+              // TODO: doesn't match browser output
+              'sizing_height_1px': '''
+Above
+
+<table border="1" style="height: 1px">
+  <tr>
+    <td style="height: 1px">Foo</td>
+  </tr>
+</table>
+
+Below''',
+              'sizing_width_100_percent': '''
+<table border="1" style="width: 100%">
+  <tr>
+    <td>One</td>
+    <td>Two</td>
+    <td>Three</td>
+  </tr>
 </table>''',
               'valign_baseline_1a': '''
 <table border="1">
@@ -1214,8 +1081,9 @@ Future<void> main() async {
     <td valign="baseline">Foo</td>
   </tr>
 </table>''',
+              // https://github.com/daohoangson/flutter_widget_from_html/issues/171
+              // https://github.com/daohoangson/flutter_widget_from_html/issues/1028
               'row_color': '''
-<!-- https://github.com/daohoangson/flutter_widget_from_html/issues/1028 -->
 <table style="border-collapse: collapse;">
   <tr>
     <th>First Name</th>
@@ -1235,7 +1103,7 @@ Future<void> main() async {
   <tr style="background-color: #f2f2f2;">
     <td>Adam</td>
     <td>Johnson</td>
-    <td>67</td>
+    <td style="background-color: red;">67</td>
   </tr>
 </table>''',
               'rtl': '''
@@ -1282,6 +1150,18 @@ Future<void> main() async {
     <td>$multiline</td>
   </tr>
 </table>''',
+              // https://github.com/daohoangson/flutter_widget_from_html/issues/1322
+              // https://github.com/daohoangson/flutter_widget_from_html/issues/1446
+              'text_align_center': '''
+<table border="1">
+  <tr>
+    <td>Long long long text</td>
+  </tr>
+  <tr>
+    <td style="text-align: center;">Short text</td>
+  </tr>
+</table>
+''',
               'width_redistribution_wide': '''
 <div style="background: red; width: 400px">
   <table border="1">
@@ -1302,6 +1182,7 @@ Future<void> main() async {
     </tr>
   </table>
 </div>''',
+              // TODO: doesn't match browser output
               'width_in_percent': '''
 <table border="1">
   <tr>
@@ -1528,8 +1409,8 @@ Future<String> explain(
 
 String _padding(String child) => '[HtmlTableCell:child='
     '[Padding:(1,1,1,1),child='
-    '[Align:alignment=centerLeft,widthFactor=1.0,child='
-    '$child]]]';
+    '[Align:alignment=centerLeft,widthFactor=1.0,child=[CssBlock:child='
+    '$child]]]]';
 
 final _loggerIsGitHubAction = Platform.environment['GITHUB_ACTIONS'] == 'true';
 final _loggerMessages = [];
@@ -1566,7 +1447,7 @@ class _Golden extends StatelessWidget {
   const _Golden(this.contents);
 
   @override
-  Widget build(BuildContext _) => Scaffold(
+  Widget build(BuildContext context) => Scaffold(
         body: Padding(
           padding: const EdgeInsets.all(8.0),
           child: HtmlWidget(contents),
