@@ -27,7 +27,7 @@ Future<String> explain(
   WidgetTester tester,
   String html, {
   bool useExplainer = true,
-}) async =>
+}) =>
     helper.explain(
       tester,
       null,
