@@ -4,5 +4,5 @@ import 'web_view.dart';
 
 class WebViewState extends State<WebView> {
   @override
-  Widget build(BuildContext _) => throw UnimplementedError();
+  Widget build(BuildContext context) => throw UnimplementedError();
 }
